add util.EncodeUUIDHex
