+++ pkg/util/export/table/type.go
@@ -16,7 +16,6 @@ package table
 
 import (
 	"context"
-	"encoding/hex"
 	"fmt"
 	"io"
 	"path"
@@ -288,20 +287,6 @@ func GetExtension(ext string) string {
 	default:
 		panic("unknown type of ext")
 	}
-}
-
-// EncodeUUIDHex encode uuid to string
-// len(dst) >= 36, and uuid must be [16]byte
-func EncodeUUIDHex(dst []byte, uuid []byte) {
-	hex.Encode(dst, uuid[:4])
-	dst[8] = '-'
-	hex.Encode(dst[9:13], uuid[4:6])
-	dst[13] = '-'
-	hex.Encode(dst[14:18], uuid[6:8])
-	dst[18] = '-'
-	hex.Encode(dst[19:23], uuid[8:10])
-	dst[23] = '-'
-	hex.Encode(dst[24:], uuid[10:16])
 }
 
 type RowWriter interface {
